+++ src/plugins/extra/linkbrowser/lib/linkbrowser-plugin.js
@@ -50,7 +50,24 @@ define( [
 		
 		onSelect: function ( item ) {
 			Links.hrefField.setItem( item );
-			Links.hrefChange();
+			
+			// Now create a selection within the editable since the user should be able to type once the link has been created.
+			// 1. We need to save the current cursor position since the a activate editable event will be fired and this will 
+			// set the cursor in the upper left cornor of the editable.
+			var	range = Aloha.Selection.getRangeObject();
+			var currentStartContainer = range.startContainer = range.endContainer; 
+			var currentStartOffset = range.startOffset = range.endOffset;
+			
+			// 2. Do the first select - this will invoke the activate editable event
+			range.select();
+
+			// 3. Restore the range
+			range.startContainer = range.endContainer = currentStartContainer; 
+			range.startOffset = range.endOffset = currentStartOffset;
+			// 4. Invoke the final selection
+			range.select();
+						
+			// Close the browser lightbox
 			this.close();
 		},